refactor: split training.py into training/ subpackage + extract eval_loop
training/
├── __init__.py    # public re-exports
├── result.py      # TrainingResult dataclass
├── schedules.py   # cosine_with_warmup (was _cosine_with_warmup)
├── single.py      # train_basis
├── batched.py     # train_basis_batched orchestrator (uses eval_loop)
├── adam_step.py   # _build_jit_adam_step (JIT closure factory)
└── eval_loop.py   # NEW: evaluate_and_check_early_stop

DRY fix: the early-stopping + validation bookkeeping previously
duplicated in both Adam and GD branches of train_basis_batched
(lines 514-533 vs 584-599 of training.py) is now one helper called
from both paths. Behavior is byte-exact: SHA-256 of loss_history on
the QFT(2,2) seed=0 / 20-step fixture matches pre-refactor.

The Adam JIT step (training/adam_step.py) intentionally remains
distinct from optimizers/adam.py (different representation for
XLA-friendly compilation).

Parity tests pass: test_parity_training, test_parity_adam, test_parity_long_run.
