+++ tests/training/test_single.py
